+++ src/create-tests/config/Configuration.ts
@@ -1,20 +1,12 @@
-import * as fs from "fs";
 import * as vscode from "vscode";
-import * as path from "path";
 import { WorkspaceConfiguration } from "vscode";
 import { IConfiguration, TestTask } from "../types";
 import { SourceFile } from "../SourceFile";
-import {
-  getDirectoryPath,
-  replaceSourceDir,
-  getTestFileName,
-  isTestDirectory,
-} from "../utils";
+
 import {
   DEFAULT_TEST_FILES_SUFFIX,
   DefaultLocationForTestFiles,
 } from "../constants";
-import mkdirp = require("mkdirp");
 
 /**
  * A class used to hold the extension configurations. It also defines the getters to read the configuration.
@@ -33,11 +25,11 @@ export class Configuration {
   };
 
   private workspaceConfig: WorkspaceConfiguration;
-  private sourceFile: SourceFile;
+  // private sourceFile: SourceFile;
 
   constructor(config: WorkspaceConfiguration, sourceFile: SourceFile) {
     this.workspaceConfig = config;
-    this.sourceFile = sourceFile;
+    // this.sourceFile = sourceFile;
   }
 
   /**
@@ -66,84 +58,6 @@ export class Configuration {
 
   getCustomLocationForTestFiles(): string {
     return this.getConfigValue("customLocationForFiles", "");
-  }
-
-  /**
-   * This method reads the configuration and based on that returns the path of directory inside which test file
-   * should be created. First it checks whether the `customLocationForFiles` is set or not. If the configuration
-   * is absent then it reads the `defaultLocationForFiles` and based on that returns the directory path.
-   */
-  getTestFilesLocation(): string {
-    let filesLocation: string = this.getCustomLocationForTestFiles();
-
-    if (this.isTestFile(this.sourceFile)) {
-      return this.sourceFile.getDirectoryPath();
-    }
-
-    if (filesLocation) {
-      // If the path given by user is not absolute then first make it absolute
-      if (!path.isAbsolute(filesLocation)) {
-        filesLocation = path.normalize(
-          path.join(this.sourceFile.getBaseDirectoryPath(), filesLocation)
-        );
-      }
-
-      const sourceFilePath: string = this.sourceFile.getRelativeFileDirname();
-
-      if (sourceFilePath.indexOf(path.sep) !== -1) {
-        filesLocation = path.join(
-          filesLocation,
-          getDirectoryPath(sourceFilePath)
-        );
-      }
-
-      filesLocation = replaceSourceDir(filesLocation, this.getSourceDir());
-
-      return filesLocation;
-    }
-
-    switch (this.getDefaultLocationForTestFiles()) {
-      case DefaultLocationForTestFiles.SAME_AS_SOURCE_FILE:
-        let localFilePath: string = this.sourceFile.getDirectoryPath();
-
-        if (
-          !this.isTestFile(this.sourceFile) &&
-          isTestDirectory(this.getTestDirectoryName(), localFilePath)
-        ) {
-          localFilePath = path.join(localFilePath, this.getTestDirectoryName());
-        }
-
-        return localFilePath;
-
-      case DefaultLocationForTestFiles.PROJECT_ROOT:
-        const sourceFilePath: string = this.sourceFile.getRelativeFileDirname();
-
-        // Check whether the SourceFile is present inside multiple directories
-        return sourceFilePath.indexOf(path.sep) > -1
-          ? this.mimicSourceDirectoryStructure()
-          : this.sourceFile.getBaseDirectoryPath();
-    }
-
-    return "";
-  }
-
-  /**
-   * This method is used to mimic the directory structure of the source file.
-   */
-  private mimicSourceDirectoryStructure(): string {
-    let directoryPath: string = path.join(
-      this.sourceFile.getBaseDirectoryPath(),
-      this.getTestDirectoryName(),
-      getDirectoryPath(this.sourceFile.getRelativeFileDirname())
-    );
-
-    directoryPath = replaceSourceDir(directoryPath, this.getSourceDir());
-
-    if (!fs.existsSync(directoryPath)) {
-      mkdirp.sync(directoryPath);
-    }
-
-    return directoryPath;
   }
 
   getTestDirectoryName(): string {
@@ -198,57 +112,13 @@ export class Configuration {
       this.defaultConfiguration.supportedExtension
     );
   }
+
   isValidExtension(sourceFile: SourceFile) {
     const ext = sourceFile.getExtension();
     const extensions = this.getSupportedExtension();
     return extensions.indexOf(ext) !== -1;
   }
-  isTestFile(sourceFile: SourceFile) {
-    const ext = this.getTestFilesSuffix() + "." + sourceFile.getExtension();
-    return sourceFile.getName().endsWith(ext);
-  }
 
-  getParentFileName(sourceFile: SourceFile) {
-    const fileName = sourceFile.getNameWithoutExtension();
-    const originalFile = fileName.replace("." + this.getTestFilesSuffix(), "");
-    return originalFile + "." + sourceFile.getExtension();
-    // const parentFile =  this.sourceFile.getName().;
-  }
-  getParentSourceFile(sourceFile: SourceFile) {
-    const parentDir = sourceFile
-      .getDirectoryPath()
-      .replace(this.getTestDirectoryName(), "");
-    const filePath = path.join(parentDir, this.getParentFileName(sourceFile));
-    return new SourceFile(vscode.Uri.file(filePath));
-  }
-  getTestFilesDirectory() {
-    let testDirPath = this.getTestFilesLocation();
-
-    if (isTestDirectory(this.getTestDirectoryName(), testDirPath)) {
-      const testDirName = this.getTestDirectoryName();
-      testDirPath = path.join(testDirPath, testDirName);
-    }
-    return testDirPath;
-  }
-  getTestSourceFile() {
-    return new SourceFile(vscode.Uri.file(this.getTestFileAbsolutePath()));
-  }
-
-  getTestFileAbsolutePath() {
-    let testDirPath = this.getTestFilesLocation();
-
-    if (!isTestDirectory(this.getTestDirectoryName(), testDirPath)) {
-      const testDirName = this.getTestDirectoryName();
-      testDirPath = path.join(testDirPath, testDirName);
-    }
-
-    const fileName = this.isTestFile(this.sourceFile)
-      ? this.sourceFile.getName()
-      : getTestFileName(this.sourceFile, this);
-    const testFilePath = path.join(testDirPath, fileName);
-
-    return testFilePath;
-  }
   getTask(taskName: string) {
     const tasks = this.getTasks();
     let task = tasks.find((x) => x.label === taskName);
